% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dat.R
\name{dat_to_plot}
\alias{dat_to_plot}
\title{Format data for plotting}
\usage{
dat_to_plot(raw_dat)
}
\arguments{
\item{raw_dat}{Input data with columns x,x_err,y,y_err}
}
\value{
Data formatted for plotting with measurement errors
}
\description{
Format data for plotting
}
