% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NJ_CC}
\alias{NJ_CC}
\title{A sea level reconstruction from New Jersey, USA.}
\format{
A data frame with 135 rows and 4 variables:
\describe{
  \item{x}{age, in year CE}
  \item{y}{relative sea level in metres}
  \item{x_err}{age error}
  \item{y_err}{RSL error}
}
}
\source{
\url{https://cp.copernicus.org/articles/12/525/2016/}
}
\usage{
NJ_CC
}
\description{
A dataset containing the relative sea level (RSL) reconstruction, the RSL error, the age reconstruction and the agg error
}
\keyword{datasets}
