% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{par_est}
\alias{par_est}
\title{Get parameter estimates and model estimates}
\usage{
par_est(mod)
}
\arguments{
\item{mod}{Fitted model object}
}
\value{
List with model estimates (pred_res) and parameter estimates (par_summary)
}
\description{
Get parameter estimates and model estimates
}
\examples{
dat <- sim_slr(n_sim = 30)
mod <- run_mod(dat, model = "slr")
par_est(mod)

}
