% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dat.R
\name{plot_dat}
\alias{plot_dat}
\title{Plot data with measurement uncertainty (in y and x variables)}
\usage{
plot_dat(dat, BP_scale = FALSE, add_truth = FALSE)
}
\arguments{
\item{dat}{Input data with columns x,x_err,y,y_err}

\item{BP_scale}{Present the data as Before Present (BP). Defaults to FALSE.}

\item{add_truth}{Logical argument to add the "True" data generating process to the plot. This should only be set to True when using simulated data from sim_slr, sim_cp or sim_gp.}
}
\value{
Plot of data with measurement errors
}
\description{
Plot data with measurement uncertainty (in y and x variables)
}
\examples{
dat <- sim_slr(n_sim = 30)
plot_dat(dat)
}
