% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{run_mod}
\alias{run_mod}
\title{Fit errors-in-variables models with JAGS}
\usage{
run_mod(
  dat,
  model = "slr",
  EIV = TRUE,
  BP_scale = FALSE,
  n_cp = 1,
  igp_smooth = 0.2,
  iter = 5000,
  burnin = 1000,
  thin = 4,
  scale_factor_x = 1,
  scale_factor_y = 1
)
}
\arguments{
\item{dat}{Input data with columns x,x_err,y,y_err}

\item{model}{The model to run. Choose from slr, cp, gp, igp}

\item{EIV}{Use EIV framework. Defaults to TRUE}

\item{BP_scale}{Present the data as Before Present (BP). Defaults to FALSE.}

\item{n_cp}{Number of change points if model = "cp" is chosen. Can choose from 1,2,3,4.}

\item{igp_smooth}{Informs prior for the smoothness (correlation) parameter if model = "igp" is chosen. Choose a value between 0 and 1. Closer to 1 will increase smoothness.}

\item{iter}{MCMC iterations}

\item{burnin}{MCMC burnin}

\item{thin}{MCMC thinning}

\item{scale_factor_x}{value to divide the predictor (x) by to change the scale. 1 will result in no change. 1000 is recommended if x is years.}

\item{scale_factor_y}{value to divide the response (y) by to change the scale. 1 will result in no change.}
}
\value{
List with data, JAGS input data, model output and the name of the model file.
}
\description{
Fit errors-in-variables models with JAGS
}
\examples{
dat <- sim_slr(n_sim = 30)
mod <- run_mod(dat, model = "slr")
}
