% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_slr}
\alias{sim_slr}
\title{Simulate data with measurment error from a simple linear regression}
\usage{
sim_slr(
  n_sim = 50,
  min_x = 0,
  max_x = 2,
  alpha = 0,
  beta = 1,
  sigma = 0.1,
  x_err = 0.1,
  y_err = 0.1
)
}
\arguments{
\item{n_sim}{Number of data points to simulate}

\item{min_x}{Minimum x value}

\item{max_x}{Maximum x value}

\item{alpha}{Regression intercept}

\item{beta}{Regression slope}

\item{sigma}{Nugget variation}

\item{x_err}{x measurement error}

\item{y_err}{y measurement error}
}
\value{
Simulated dataset with columns x, x_err, y, y_err
}
\description{
Simulate data with measurment error from a simple linear regression
}
\examples{
sim_slr(n_sim = 50)
}
