% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{run_mod}
\alias{run_mod}
\title{Fit errors-in-variables models with JAGS}
\usage{
run_mod(
  dat,
  model = "model_eiv_reg",
  iter = 15000,
  burnin = 5000,
  thin = 5,
  scale_factor = 1
)
}
\arguments{
\item{dat}{Input data with columns x,x_err,y,y_err}

\item{model}{The model to run. Choose from model_reg, model_eiv_reg, model_cp, model_eiv_cp, model_gp, model_eiv_gp, model_eiv_igp. Defaults to model_eiv_reg.}

\item{iter}{MCMC iterations}

\item{burnin}{MCMC burnin}

\item{thin}{MCMC thinning}

\item{scale_factor}{value to divide the predictor (x) by to change the scale. 1 will result in n0 change. 1000 is recommended if x is years.}
}
\value{
List with data, JAGS input data, model output and the name of the model file.
}
\description{
Fit errors-in-variables models with JAGS
}
\examples{
dat <- sim_reg(n_sim = 30)
mod <- run_mod(dat, model = "model_eiv_reg")
}
