% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{run_mod}
\alias{run_mod}
\title{Fit errors-in-variables models with JAGS}
\usage{
run_mod(
  dat,
  model = "slr",
  EIV = TRUE,
  iter = 5000,
  burnin = 1000,
  thin = 4,
  scale_factor = 1
)
}
\arguments{
\item{dat}{Input data with columns x,x_err,y,y_err}

\item{model}{The model to run. Choose from slr, cp, gp, igp}

\item{EIV}{Indicate whether to use EIV framework. Defaults to TRUE}

\item{iter}{MCMC iterations}

\item{burnin}{MCMC burnin}

\item{thin}{MCMC thinning}

\item{scale_factor}{value to divide the predictor (x) by to change the scale. 1 will result in n0 change. 1000 is recommended if x is years.}
}
\value{
List with data, JAGS input data, model output and the name of the model file.
}
\description{
Fit errors-in-variables models with JAGS
}
\examples{
dat <- sim_slr(n_sim = 30)
mod <- run_mod(dat, model = "slr")
}
